feat : promotion 별도 도메인으로 분리
