+++ src/main/java/com/avg/lawsuitmanagement/promotion/repository/PromotionMapperRepository.java
@@ -1,0 +1,9 @@
+package com.avg.lawsuitmanagement.promotion.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface PromotionMapperRepository {
+
+    void insertPromotionKey(String value);
+}
