+++ src/main/resources/mybatis/mapper/Promotion.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.avg.lawsuitmanagement.promotion.repository.PromotionMapperRepository">
+
+    <insert id="insertPromotionKey" parameterType="java.lang.String">
+        replace into client_promotion (`value`)
+        values (#{value})
+    </insert>
+
+</mapper>
